pyproject.toml
[project]
name = "amqcsldb-py"
version = "1.0.6"
description = "Add your description here"
readme = "README.rst"
authors = [
    { name = "stickie", email = "71910589+FieryIceStickie@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "attrs>=25.3.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "typer>=0.15.4",
]

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "single"

[tool.pyright]
typeCheckingMode = "strict"
reportUnusedCallResult = false
reportUnusedVariable = false
reportUnreachable = false
reportUnnecessaryComparison = false

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = [
    "ruff>=0.11.7",
]

[project.scripts]
amqcsl = "amqcsl.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/amqcsl"]
